Clippy modifications
These changes are motivated by lints suggested by clippy. The changes
inside the benches/insert _do_ fix compilation bugs which, uh, shows
we aren't running our benchmarks enough.

+++ benches/insert.rs
@@ -9,7 +9,7 @@ use quantiles::ckms::CKMS;
 #[bench]
 fn bench_sequential_insert(b: &mut Bencher) {
     b.iter(|| {
-               let mut ckms = CKMS::<i64>::new(0.001);
+               let mut ckms = CKMS::<i32>::new(0.001);
                for i in 0..10000 {
                    ckms.insert(i);
                }
@@ -21,7 +21,7 @@ fn bench_sequential_insert(b: &mut Bencher) {
 fn bench_inverted_insert(b: &mut Bencher) {
     b.iter(|| {
                let seq = (0..10000).rev();
-               let mut ckms = CKMS::<i64>::new(0.001);
+               let mut ckms = CKMS::<i32>::new(0.001);
                for i in seq {
                    ckms.insert(i);
                }
